+++ src/main/java/com/app/domain/categorizedsummary/service/CategorizedSummaryService.java
@@ -6,12 +6,12 @@ import com.app.domain.category.entity.Category;
 import com.app.domain.category.service.CategoryService;
 import com.app.domain.member.entity.Member;
 import com.app.domain.member.service.MemberService;
-import com.app.domain.membersavedsummary.dto.MemberSavedSummaryDto;
-import com.app.domain.membersavedsummary.entity.MemberSavedSummary;
-import com.app.domain.membersavedsummary.mapper.MemberSavedSummaryMapper;
-import com.app.domain.membersavedsummary.service.MemberSavedSummaryService;
-import com.app.domain.summary.entity.AiGeneratedSummary;
-import com.app.domain.summary.service.SummaryService;
+import com.app.domain.summary.membersavedsummary.dto.MemberSavedSummaryDto;
+import com.app.domain.summary.membersavedsummary.entity.MemberSavedSummary;
+import com.app.domain.summary.membersavedsummary.mapper.MemberSavedSummaryMapper;
+import com.app.domain.summary.membersavedsummary.service.MemberSavedSummaryService;
+import com.app.domain.summary.aigeneratedsummary.entity.AiGeneratedSummary;
+import com.app.domain.summary.aigeneratedsummary.service.SummaryService;
 import com.app.global.error.ErrorCode;
 import com.app.global.error.exception.BusinessException;
 import com.app.global.error.exception.EntityNotFoundException;
