+++ src/main/java/com/app/domain/summary/aigeneratedsummary/repository/AiGeneratedSummaryRepository.java
@@ -1,0 +1,13 @@
+package com.app.domain.summary.aigeneratedsummary.repository;
+
+import com.app.domain.summary.aigeneratedsummary.entity.AiGeneratedSummary;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface AiGeneratedSummaryRepository extends JpaRepository<AiGeneratedSummary,Integer> {
+
+    Optional<AiGeneratedSummary> findBySummaryFile_FileId(int fileId);
+}
