--- src/main/java/com/app/domain/summary/dto/SummaryFile/AiRequest/AiGenerateSummaryFromAiDto.java
@@ -1,14 +1,0 @@
-package com.app.domain.summary.dto.SummaryFile.AiRequest;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Setter
-@Getter
-public class AiGenerateSummaryFromAiDto {
-    private String summaryContent;
-}
