+++ src/main/java/com/app/global/pdf/SummaryPdfMaker.java
@@ -1,6 +1,6 @@
 package com.app.global.pdf;
 
-import com.app.domain.summary.dto.SummaryFile.AiRequest.AiGenerateSummaryFromAiDto;
+import com.app.domain.summary.aigeneratedsummary.dto.SummaryFile.AiRequest.AiGenerateSummaryFromAiDto;
 import com.app.global.config.ENUM.PdfType;
 import org.apache.pdfbox.pdmodel.PDDocument;
 import org.apache.pdfbox.pdmodel.PDPage;
